Actualizacion de importacion productos & add logos Niubiz & styles marketplace

+++ database/migrations/2023_06_14_170731_create_category_subcategory_table.php
@@ -15,8 +15,8 @@ class CreateCategorySubcategoryTable extends Migration
     {
         Schema::create('category_subcategory', function (Blueprint $table) {
             $table->smallIncrements('id')->unsigned();
-            $table->unsignedTinyInteger('category_id')->nullable();
-            $table->unsignedSmallInteger('subcategory_id')->nullable();
+            $table->unsignedInteger('category_id')->nullable();
+            $table->unsignedInteger('subcategory_id')->nullable();
             $table->foreign('category_id')->on('categories')->references('id');
             $table->foreign('subcategory_id')->on('subcategories')->references('id');
         });
